+++ ExpandWorldEvents/EventData.cs
@@ -1,4 +1,6 @@
+using System.Collections.Generic;
 using System.ComponentModel;
+using ExpandWorldData;
 namespace ExpandWorld.Event;
 
 public class Data
@@ -8,8 +10,8 @@ public class Data
   public bool enabled = true;
   [DefaultValue(60f)]
   public float duration = 60f;
-  [DefaultValue(true)]
-  public bool nearBaseOnly = true;
+  [DefaultValue("3")]
+  public string nearBaseOnly = "3";
   [DefaultValue("")]
   public string biome = "";
   [DefaultValue("")]
@@ -38,4 +40,25 @@ public class Data
   public bool pauseIfNoPlayerInArea = true;
   [DefaultValue(true)]
   public bool random = true;
+  [DefaultValue(100f)]
+  public float playerDistance = 100f;
+  [DefaultValue("")]
+  public string playerLimit = "";
+}
+public class ExtraData
+{
+  public List<string> RequiredEnvironments = [];
+  public float PlayerDistance = 100f;
+  public Range<int>? PlayerLimit;
+  public int BaseValue = 3;
+
+  public ExtraData(Data data)
+  {
+    RequiredEnvironments = DataManager.ToList(data.requiredEnvironments);
+    PlayerDistance = data.playerDistance;
+    if (data.playerLimit != "")
+      PlayerLimit = Parse.IntRange(data.playerLimit);
+    if (data.nearBaseOnly != "true")
+      BaseValue = Parse.Int(data.nearBaseOnly, 0);
+  }
 }