+++ ExpandWorldEvents/ExpandWorldEvents.cs
@@ -9,12 +9,12 @@ using Service;
 using UnityEngine;
 namespace ExpandWorld;
 [BepInPlugin(GUID, NAME, VERSION)]
-[BepInDependency("expand_world_data", "1.21")]
+[BepInDependency("expand_world_data", "1.22")]
 public class EWE : BaseUnityPlugin
 {
   public const string GUID = "expand_world_events";
   public const string NAME = "Expand World Events";
-  public const string VERSION = "1.4";
+  public const string VERSION = "1.5";
 #nullable disable
   public static ManualLogSource Log;
 #nullable enable